Extarct CALLBACK_TYPE
