+++ packages/react-native-gesture-handler/src/handlers/gestures/gesture.ts
@@ -1,3 +1,4 @@
+import { CALLBACK_TYPE } from '../../CallbackType';
 import { isRemoteDebuggingEnabled } from '../../utils';
 import type {
   ActiveCursor,
@@ -90,24 +91,6 @@ export type HandlerCallbacks<EventPayloadT extends Record<string, unknown>> = {
   ) => GestureUpdateEvent<Record<string, unknown>>;
   isWorklet: boolean[];
 };
-
-export const CALLBACK_TYPE = {
-  UNDEFINED: 0,
-  BEGAN: 1,
-  START: 2,
-  UPDATE: 3,
-  CHANGE: 4,
-  END: 5,
-  FINALIZE: 6,
-  TOUCHES_DOWN: 7,
-  TOUCHES_MOVE: 8,
-  TOUCHES_UP: 9,
-  TOUCHES_CANCEL: 10,
-} as const;
-
-// Allow using CALLBACK_TYPE as object and type
-// eslint-disable-next-line @typescript-eslint/no-redeclare
-export type CALLBACK_TYPE = (typeof CALLBACK_TYPE)[keyof typeof CALLBACK_TYPE];
 
 export abstract class Gesture {
   /**
